Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -16,11 +16,11 @@ int main(void)
 	long long ll;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (sizeof(c)));
-	puts("Size of an int: %lu byte(s)", sizeof(i));
-	puts("Size of a long int: %lu byte(s)", sizeof(l));
-	puts("Size of a long long int: %lu byte(s)", sizeof(ll));
-	puts("Size of a float: %lu byte (s)", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	puts("Size of an int: %lu byte(s)", (unsigned long)sizeof(i));
+	puts("Size of a long int: %lu byte(s)", (unsigned long)sizeof(l));
+	puts("Size of a long long int: %lu byte(s)", (unsigned long)sizeof(ll));
+	puts("Size of a float: %lu byte (s)", (unsigned long)sizeof(f));
 
 	return (0);
 }
